#280 #16 Moved de/serialization of primitives into same file
